{{cookiecutter.repo_name}}/requirements.txt
pandas
sacred
sklearn
scipy
PyYAML
munch
pyarrow
mltome==0.1.0
dask-ml
joblib
mlflow
fire
